Manages plotting for PROAST

Runs through the plotting functions depending on data type and plot type.

Usage

f.plot.gui(
  ans.all,
  HTML = FALSE,
  model.summ = TRUE,
  display_plots = TRUE,
  .proast_env = NULL,
  output_type = NULL,
  filename = NULL,
  interactive_mode = TRUE,
  record_plots = FALSE
)

Arguments

ans.all

The proast object that gets passed to all functions.

HTML

Keep FALSE

model.summ

Keep TRUE

display_plots

A logical variable - whether we want to display the plots or not.

.proast_env

environment

output_type

The format that you wish to export the plots as.

filename

The name of the file to be read.

interactive_mode

A TRUE/FALSE value specifying whether you want to run interactively (i.e., TRUE, the default) or using command-line mode (i.e., FALSE, non-interactive). If FALSE, you must provide all other parameters.

record_plots

A logical variable indicating whether you want to record the plots and return them as a list instead of exporting them. This parameter should only be used when running the function indenpently or within f.plot.result. It will disrupt f.proast(), so keep the default as FALSE.