Write a Mutational Matrix to input into the sigprofiler web application

Creates a .txt file from mutation data that can be used for mutational signatures analysis using the SigProfiler web application. Can handle group analyses (ex dose, tissue, etc). Currently only supports SBS matrices i.e. snvs.

Usage

write_mutational_matrix(
  mutation_data,
  group = "dose",
  subtype_resolution = "base_96",
  mf_type = "min",
  output_path = NULL
)

Arguments

mutation_data

The object containing the mutation data. The output of import_mut_data() or import_vcf_data().

group

The column in the mutation data used to aggregate groups (e.g., sample, tissue, dose).

subtype_resolution

The resolution of the mutation subtypes. Options are "base_6" or "base_96". Default is "base_96".

mf_type

The mutation counting method to use. Options are "min" or "max". Default is "min".

output_path

The path to save the output file. If not provided, the file will be saved in the current working directory. Default is NULL.

Value

a .txt file that can be uploaded to the SigProfiler Assignment web application (https://cancer.sanger.ac.uk/signatures/assignment/) as a "Mutational Matrix".

Details

Mutations will be be filtered for SNVs. Mutations flagged in filter_mut will be excluded from the output. Mutations will be summed across the groups specified in the group argument.

Examples

example_file <- system.file("extdata", "Example_files",
                            "example_mutation_data_filtered.rds",
                            package = "MutSeqR")
example_data <- readRDS(example_file)
temp_output <- tempdir()
write_mutational_matrix(mutation_data = example_data,
                        group = "dose_group",
                        subtype_resolution = "base_96",
                        mf_type = "min",
                        output_path = temp_output)
list.files(temp_output)
#>  [1] "bslib-2ce9eb5d8d7866b4f2d82a9ef19eb08f"                                                        
#>  [2] "dose_group_base_96_mutational_matrix.txt"                                                      
#>  [3] "downlit"                                                                                       
#>  [4] "file32db20eb7fb3"                                                                              
#>  [5] "file32db2ca5e076"                                                                              
#>  [6] "file32db555bd8ba"                                                                              
#>  [7] "file32db7556ed6e"                                                                              
#>  [8] "file32dbcc4b657"                                                                               
#>  [9] "libloc_177_9a667201821007d.rds"                                                                
#> [10] "libloc_182_ba3851d8a9100920.rds"                                                               
#> [11] "libloc_182_d3198563f59e5c50.rds"                                                               
#> [12] "mutation_calling_file.txt"                                                                     
#> [13] "repos_https%3A%2F%2Fbioconductor.org%2Fpackages%2F3.21%2Fdata%2Fannotation%2Fsrc%2Fcontrib.rds"
# The file is saved in the temporary directory
# To view the file, use the following code:
## output_file <- file.path(temp_output, "dose_group_base_96_mutational_matrix.txt")
## file.show(output_file)